proj1/requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.38
pydantic>=2.11.1
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.13.3
lxml>=5.3.1
playwright>=1.52.0
